requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
python-telegram-bot==13.2
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
yapf==0.30.0
